Förenklat filtreringsfunktioner med comprehensions och flyttat git-hash-test till tests/
